+++ nevergrad/optimization/multiobjective/test_core.py
@@ -3,8 +3,14 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
+import warnings
+import typing as tp
+from unittest import SkipTest
 import numpy as np
+import pytest
 import nevergrad as ng
+from .. import base
+from ..optimizerlib import registry
 from . import core
 
 
@@ -30,7 +36,31 @@ def test_hypervolume_pareto_function() -> None:
     assert front == expected_front, f"Expected {expected_front} but got {front}"
 
 
-# pylint: disable=redefined-outer-name,unsubscriptable-object,unused-variable,unused-import,reimported,import-outside-toplevel
+@pytest.mark.parametrize("losses", [(12, [3, 4]), ([3, 4], 12)])  # type: ignore
+def test_num_losses_error(losses: tp.Tuple[tp.Any, tp.Any]) -> None:
+    opt = ng.optimizers.CMA(parametrization=3, budget=100)
+    cand = opt.ask()
+    opt.tell(cand, losses[0])
+    cand = opt.ask()
+    with pytest.raises(ValueError):
+        opt.tell(cand, losses[1])
+
+
+@pytest.mark.parametrize("name", registry)  # type: ignore
+def test_optimizers_multiobjective(name: str) -> None:  # pylint: disable=redefined-outer-name
+    if "BO" in name:
+        raise SkipTest("BO is currently failing for unclear reasons")  # TODO solve
+    with warnings.catch_warnings():
+        # tests do not need to be efficient
+        warnings.simplefilter("ignore", category=base.InefficientSettingsWarning)
+        optimizer = registry[name](parametrization=4, budget=100)
+        candidate = optimizer.ask()
+        optimizer.tell(candidate, [12, 12])
+        optimizer.ask()
+
+
+# pylint: disable=redefined-outer-name,unsubscriptable-object,
+# pylint: disable=unused-variable,unused-import,reimported,import-outside-toplevel
 def test_doc_multiobjective() -> None:
     # DOC_MULTIOBJ_OPT_0
     import nevergrad as ng
@@ -39,32 +69,30 @@ def test_doc_multiobjective() -> None:
     def multiobjective(x):
         return [np.sum(x**2), np.sum((x - 1)**2)]
 
-    # upper_bounds=[2.5, 2.5])
-
     print("Example: ", multiobjective(np.array([1.0, 2.0, 0])))
     # >>> Example: [5.0, 2.0]
 
-    optimizer = ng.optimizers.CMA(parametrization=3, budget=100)  # 3 is the dimension, 100 is the budget.
-    # optimizer = ng.optimizers.OnePlusOne(parametrization=3, budget=100)  # 3 is the dimension, 100 is the budget.
+    optimizer = ng.optimizers.CMA(parametrization=3, budget=100)
 
-    # it's not strictly necessary but highly advised to provide an upper bound reference for the losses
-    # (if not provided, such upper bound is automatically computed with the first few "tell")
+    # it's not strictly necessary but highly advised to provide an
+    # upper bound reference for the losses (if not provided, such upper
+    # bound is automatically inferred with the first few "tell")
     optimizer.tell(ng.p.MultiobjectiveReference(), [5, 5])
-    # note that you can provide a Parameter to MultiobjectiveReference, which will be passed to the
-    # optimizer
+    # note that you can provide a Parameter to MultiobjectiveReference,
+    # which will be passed to the optimizer
 
     optimizer.minimize(multiobjective, verbosity=2)
 
     # The function embeds its Pareto-front:
     print("Pareto front:")
-    for param in sorted(optimizer.pareto_front(), key=lambda p: p.losses[0]):  # type: ignore
+    for param in sorted(optimizer.pareto_front(), key=lambda p: p.losses[0]):
         print(f"{param} with losses {param.losses}")
 
     # >>> Array{(3,)}:[0. 0. 0.] with loss [0. 3.]
     #     Array{(3,)}:[0.39480968 0.98105712 0.55785803] with loss [1.42955333 0.56210368]
     #     Array{(3,)}:[1.09901515 0.97673712 0.97153943] with loss [3.10573857 0.01115516]
 
-    # It can also provide a subset:
+    # It can also provide subsets:
     print("Random subset:", optimizer.pareto_front(2, subset="random"))
     print("Loss-covering subset:", optimizer.pareto_front(2, subset="loss-covering"))
     print("Domain-covering subset:", optimizer.pareto_front(2, subset="domain-covering"))
